requirements.txt
# Backend Framework
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
gunicorn==21.2.0

# Database
pymongo==4.6.1
mongoengine==0.28.2

# ML & NLP
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
spacy==3.7.2
nltk==3.8.1

# Fairness & Bias Detection
aif360==0.5.0
fairlearn==0.10.0

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# API & Web
requests>=2.32.3
python-dotenv==1.0.0

# Email & Scheduling
apscheduler==3.10.4

# LinkedIn API (optional)
linkedin-api==2.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
